requirements.txt
supabase==2.10.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pdfplumber==0.11.4
requests==2.32.3